Replaced old player input script with new player input script
