--- Project Ronin/Assets/Scripts/Character Controller/Player_Brain_Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 273b8b85bd2a41f4ebe4705291ffb621
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
